Language bindings → Binding generators
The fact that most of these are for other languages is secondary;
the main thing they're bridging to is a different object model.
For example, GTK (GObject) is written in C, and the binding generator
has the same kind of requirements as the others.

These are *generators*. Some simpler C projects have hand-rolled
Python bindings, but those ast as normal extensions.